Move time format preference to General settings
